python/nanobot_skill/requirements.txt
python-telegram-bot==20.5
requests==2.31.0
PyYAML==6.0
